\name{dpopus}
\alias{dpopus}
\docType{data}
\title{dpopus data}
\description{simulated population of four generations}
\usage{data(dpopus)}
\format{
  A data frame with 111 age groups (single years of age).
  \describe{
    \item{\code{Females}}{Female population}
    \item{\code{Males}}{Male population}
    }
}
\source{The data are downloaded from the Human Mortality Database (HMD) and the Human Fertility Database (HFD). Country: USA. Year: 2019}
\keyword{datasets}