% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e0.R
\name{e0}
\alias{e0}
\title{Mean Ages at Death and Probabilities of Surviving to Selected Ages, by Sex}
\usage{
e0(d)
}
\arguments{
\item{d}{The name of the database. If missing, dLH is used if it exists.}
}
\value{
\item{e0}{Mean ages at death} \item{Prob65}{Probability of surviving
at age 65} \item{Prob85}{Probability of surviving at age 85}
}
\description{
Computes (a) Life expectancy at birth, (b) Probability of surviving at age
65, and (c) Probability of surviving at age 85
}
\examples{


utils::data(dLH,package="VirtualPop")
e0(d=dLH)


}
