\name{rates}
\alias{rates}
\docType{data}
\title{rates data}
\description{Mortality rates by age and sex: fertility rates by age and birth order}
\usage{data(rates)}
\format{
  A list of three objects with mortality and fertility rates.
  \describe{
    \item{\code{ASDR}}{Mortality rates}
    \item{\code{ASFR}}{Fertility rates}
    \item{\code{ratesM}}{Multistate transition rates}
    }
}
\source{The data are downloaded from the Human Mortality Database (HMD) and the Human Fertility Database (HFD). Country: USA. Year: 2019}
\keyword{datasets}
