% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualPop-package.R
\docType{data}
\name{ratesC}
\alias{ratesC}
\title{Cohort rates}
\format{
A list of three objects. \describe{
\item{ASDR}{Mortality rates} 
\item{ASFR}{Fertility rates}
\item{ratesM}{Multistate transition rates} 
}
The object returned by the function has five attributes:
\itemize{
 \item Country
 \item type: Type of data (period data or cohort data)
 \item cohort: Birth cohort (year of birth)
 \item year: Calendar year for which period death rates are used to complete cohort experience in case of incomplete mortality experience (reference year).
 \item start_pASDR: Lowest age for which cohort data are missing. The mortality rates of that age and higher ages are borrowed from period data collected in the reference year.
}
}
\source{
The data are downloaded from the Human Mortality Database (HMD) and
the Human Fertility Database (HFD). Country: USA. Cohort: 1964
}
\usage{
data(ratesC,package="VirtualPop")
}
\description{
Cohort rates of mortality by age and sex and fertility by age and parity, USA birth cohort 1964
}
