% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualPop-package.R
\docType{data}
\name{dpopus}
\alias{dpopus}
\title{dpopus data
 
Population of the United States in 2019 reported in the HMD (Population.txt file)}
\format{
A data frame with 111 age groups (single years of age). \describe{
\item{Females}{Female population}
\item{Males}{Male population}}
}
\source{
The data are downloaded from the Human Mortality Database (HMD). Country: USA. Year: 2019
}
\description{
dpopus data
 
Population of the United States in 2019 reported in the HMD (Population.txt file)
}
\keyword{datasets}
