% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Reads data from the HMD and HFD}
\usage{
GetData(country, user, pw_HMD, pw_HFD)
}
\arguments{
\item{country}{country}

\item{user}{Name of the user, used at registration with the HMD and HFD. It
is assumed that the same name is used for both HMD and HFD.}

\item{pw_HMD}{Password to access HMD, provided at registration}

\item{pw_HFD}{Password to access HFD, provided at registration}
}
\value{
\item{data_raw}{5 objects: country,life tables females,life tables males,fertility rates,female population (from HFD): exposures}
}
\description{
Reads data from the HMD and HFD
}
\examples{


\dontrun{dataLH <- GetData(country="USA",user,pw_HMD,pw_HFD)}

}
\author{
Frans Willekens
}
