% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lifespan.R
\name{Lifespan}
\alias{Lifespan}
\title{Generates Individual Lifespan(s)}
\usage{
Lifespan(data, ASDR)
}
\arguments{
\item{data}{Data frame with individual data}

\item{ASDR}{Age-specific death rates}
}
\value{
data: data frame 'dataLH' with date of death and age of death
completed.
}
\description{
Simulate length of life using age-specific death rates.Generate date of
death and age at death. The function uses the rpexp function from the
package msm and uniroot of base R
}
\examples{


utils::data(dataLH)
utils::data(rates)
z <- Lifespan (dataLH[1:5,],ASDR=rates$ASDR)


}
\author{
Frans Willekens
}
