% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartnerSearch.R
\name{PartnerSearch}
\alias{PartnerSearch}
\title{Simple Partner Search Simulation}
\usage{
PartnerSearch(dLH)
}
\arguments{
\item{dLH}{Database}
}
\value{
\item{dLH}{Updated version of database (dLH), which includes, for each individual 
without a partner and able to find a partner, the ID of the partner.}
}
\description{
In this simple model, a partner is an individual of a different sex selected at random among members of the same generation.
The function is called by GetGenerations().
}
\examples{

utils::data(dLH,package="VirtualPop")
d <- VirtualPop::PartnerSearch(dLH=dLH)

}
