% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partnership.R
\name{Partnership}
\alias{Partnership}
\title{Allocates Partners to Members of Virtual Population.}
\usage{
Partnership(dLH)
}
\arguments{
\item{dLH}{Database}
}
\value{
Updated version of database (dLH), which includes the IDs of partners.
}
\description{
Randomly allocates partners to egos
}
\examples{


utils::data(dataLH)
dLH=dataLH[1:10,]
# Remove current partner
dLH$IDpartner <- NA
d <- Partnership(dLH=dLH) 
# NOTE: partners are randomly selected from the individuals documented in dLH.

}
\author{
Frans Willekens
}
