% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualPop-package.R
\docType{data}
\name{rates}
\alias{rates}
\title{rates data}
\format{
A list of three objects. \describe{
\item{ASDR}{Mortality rates} 
\item{ASFR}{Fertility rates}
\item{ratesM}{Multistate transition rates} }
}
\source{
The data are downloaded from the Human Mortality Database (HMD) and
the Human Fertility Database (HFD). Country: USA. Year: 2019
}
\description{
Mortality rates by age and sex: fertility rates by age and birth order
}
\keyword{datasets}
